bblayers.sample file now refers to OEROOT
instead of using a hard-coded path "poky-thud", which some users
may not have, we simply use OEROOT which resolves to the poky
directory, regardless of the directory's actual name.

To use the sample files, you need to define the TEMPLATECONF
variable in your environment when sourcing the poky build
config for the first time.

$ TEMPLATECONF=meta-gateway-ww/conf source oe-init-build-env