Isolate the initramfs busybox from the rootfs one
The rootfs build of busybox has less features enabled and would
collide with other packages (e.g. sysvinit) if built the same as for
initramfs.

The quick solution is to clone the entire busybox package and rename
it, so that it can be built twice - one instance for rootfs, the other
for initramfs - each with a different set of enabled features.

Signed-off-by: Cristian Prundeanu <cristian.prundeanu@arm.com>