bring up Bluetooth hci0 on startup
In order for the Bluetooth stack to be useful, we need the
interface to be up and running after startup.

Note this script includes a workaround to Wiggle HCI at startup
time.

For some reason, when brought up during init, hci0 shows as up but
does not function. Manually (or programatically) asking it to go down
and back up solves the problem.