Verified boot: Add back in all core-os packages
When creating our original prototype of MBL secure boot with core-os we
removed some packages from the core-os image for build stability. After
rebasing on version 1.0.0 of core-os repos we can add these packages
back in with a minor change in the global-node-modules recipe: to avoid
problems with incremental builds, don't expect files in /tmp to persist
between tasks.