Allow Maestro to start without deviceos-wd
Previously, maestro.service required deviceos-wd.service. This
creates an untracked dependency between `maestro` and `deviceos-wd`.

After some experimentation, we found that maestro does not need
a running deviceos-wd to start correctly. This commit breaks this
untracked dependency and allows another development board to boot,
as it does not have a working watchdog.