coredns watcher service
Currently coredns is producing a lot of reduntant logs, as the service
is dependent on kube-bridge network interface and constantly restarts
when this is not available.

Instead create a dedicated service for starting up coredns when the
network interface becomes available.

FYI, systemd.path cannot be used as the network related files are in
/sys/.. which are not regular 'files'. Modifications do not trigger
service listeners.